chore: move fixtures to api folder
