+++ src/__tests__/pages/teachers/sitemap.test.tsx
@@ -4,7 +4,7 @@ import { GetServerSidePropsContext } from "next";
 import { getServerSideProps as getServerSidePropsSitemap } from "@/pages/teachers/sitemap.xml";
 import { getServerSideProps as getServerSidePropsSitemap1 } from "@/pages/teachers/sitemap-1.xml";
 import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
-import { teachersSitemapDataFixture } from "@/browser-lib/fixtures/teacherSitemap.fixture";
+import { teachersSitemapDataFixture } from "@/node-lib/curriculum-api-2023/fixtures/teachersSiteMap.fixture";
 
 jest.mock("next-sitemap");
